requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2